# packages: optimize list + bulk insert queries